requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.42.0
django-appconf==1.0.5
django-compressor==4.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-fontawesome-5==1.0.18
django-hitcount==1.3.5
django-model-utils==4.2.0
django-notifications==0.1.dev0
django-notifications-hq==1.7.0
idna==3.3
jsonfield==3.1.0
oauthlib==3.2.0
Pillow==9.0.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2021.5
urllib3==1.26.9
